backend/requirements.txt
fastapi==0.115.0
openai==1.59.0
python-dotenv==1.0.1
uvicorn[standard]==0.30.6
